add access control

+++ routes/articleRoute.js
@@ -1,13 +1,14 @@
 const express = require('express');
 const router = express.Router();
 const Article = require('../models/Article');
+const User = require('../models/User');
 
 /**
  * @route       /article/add
  * @method      GET
  * @description Show add article page
  */
-router.get('/add', (req, res) => {
+router.get('/add', ensureAuthenticated, (req, res) => {
     res.render('articles/add_article', {
         header: 'Add Article'
     });
@@ -19,9 +20,8 @@ router.get('/add', (req, res) => {
  * @description Add New Article
  */
 router.post('/add', (req, res) => {
-    req.checkBody('title', 'Title is required!').notEmpty(),
-    req.checkBody('author', 'Author is required!').notEmpty(),
-    req.checkBody('body', 'Body is required!').notEmpty()
+    req.checkBody('title', 'Title is required!').notEmpty();
+    req.checkBody('body', 'Body is required!').notEmpty();
 
     const errors = req.validationErrors();
     if (errors) {
@@ -33,7 +33,7 @@ router.post('/add', (req, res) => {
     else {
         let newArticle = new Article({
             title: req.body.title,
-            author: req.body.author,
+            author: req.user._id,
             body: req.body.body
         });
         newArticle.save()
@@ -50,9 +50,14 @@ router.post('/add', (req, res) => {
  * @method      GET 
  * @description Show edit article page by id
  */
-router.get('/edit/:id', (req, res) => {
+router.get('/edit/:id', ensureAuthenticated, (req, res) => {
     Article.findById(req.params.id)
         .then(article => {
+            if (article.author != req.user._id) {
+                req.flash('danger', 'Not authorized');
+                return res.redirect('/');
+            }
+
             res.render('articles/edit_article', {
                 header: 'Edit Article',
                 article
@@ -69,7 +74,6 @@ router.get('/edit/:id', (req, res) => {
 router.post('/edit/:id', (req, res) => {
     let article = {
         title: req.body.title,
-        author: req.body.author,
         body: req.body.body
     };
 
@@ -87,9 +91,22 @@ router.post('/edit/:id', (req, res) => {
  * @description Delete Article by id
  */
 router.delete('/delete/:id', (req, res) => {
-    Article.findByIdAndDelete(req.params.id)
-        .then(() => res.send('200'))
-        .catch(err => console.error(err));
+    if (!req.user._id) {
+        return res.status(500).send();
+    }
+
+    Article.findById(req.params.id)
+        .then(article => {
+            if (article.author != req.user._id) {
+                res.status(500).send();
+            }
+            else {
+                Article.findByIdAndDelete(req.params.id)
+                    .then(() => res.send('200'))
+                    .catch(err => console.error(err));
+            }
+        })
+        .catch(err => console.error(err))
 });
 
 /**
@@ -99,8 +116,26 @@ router.delete('/delete/:id', (req, res) => {
  */
 router.get('/:id', (req, res) => {
     Article.findById(req.params.id)
-        .then(article => res.render('articles/article', { article }))
+        .then(article => {
+            User.findById(article.author)
+                .then(user => res.render('articles/article', {
+                    article,
+                    author: user.name
+                }))
+                .catch(err => console.err(err));
+        })
         .catch(err => console.error(err));
 });
+
+// Access Control
+function ensureAuthenticated(req, res, next) {
+    if (req.isAuthenticated()) {
+        return next();
+    }
+    else {
+        req.flash('danger', 'Please login first');
+        res.redirect('/auth/login');
+    }
+}
 
 module.exports = router;
